feat: added notifuication when scratch org is about to expire
